Move ResourceKey to common/api. Refactor the code in reporting.
