+++ src/main/kotlin/org/wfanet/measurement/reporting/service/api.v1alpha/BUILD.bazel
@@ -11,6 +11,7 @@ kt_jvm_library(
     name = "resource_key",
     srcs = glob(["*Key.kt"]) + ["IdVariable.kt"],
     deps = [
+        "//src/main/kotlin/org/wfanet/measurement/common/api:resource_key_interface",
         "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common",
     ],
 )
